Add reclaim/uncompress ECMP groups logic
Summary:
At the end of processing routes, we may have freed
up primary ECMP groups. So after processing all routes, check
if
a. We have space to reclaim
b. We have routes pointing to back up group type.

If both are satisfied. Reclaim groups with maximum number of routes
pointing to it.
For ease of review, this function is sent as a stand alone diff.
Next diff will actually call this function and update UTs.

Reviewed By: msomasundaran

Differential Revision:
D74525225

Privacy Context Container: L1125642

fbshipit-source-id: 26e7251dd7a57cb59132e56dafd4d0d709f1692d