Enhance AgentVoqSwitchWithMultipleDsfNodesTest tests to work in multi-switch mode
Summary:
Hw_agent_tests use fb303 counters to do the tests. In mono mode, this is simply a call to wedge-agent binary. However, tests don't work in multi mode, since these counters are not available in the sw-agent.
For multi-switch mode, get the the client to hwAgent cached in SwSwitch for
 the hwAgent, and call getRegexCounters on it.

Reviewed By: msomasundaran

Differential Revision:
D74614472

Privacy Context Container: L1125642

fbshipit-source-id: 39b199cbde1fc2153213562fbc5a41234836d33d