Release SaiPhymanager in QSFP Warmboot Teardown
Summary:
The purpose of this diff is solely to address a test-only issue that arises in our QSFP hardware tests.

For warmboot shutdown, the agent handles it by leaking the hwSwitchEnsemble_ (https://fburl.com/code/4kr3gj2p) during teardown if it's a warmboot shutdown. This approach prevents the destructors from running and avoids unprogramming the hardware switches.

I confirmed with the agent team (Parvez) that this is indeed the intended behavior, and only during cold boot exits (in tests) do they go through the destructors. In Prod, destructors are not called.

In our case, we reset the ensemble regardless during the teardown phase (https://fburl.com/code/7hxo5ptb). This action attempts to remove ports and port connectors when destroying the manager table and nested port managers for Sai switches. However, by this stage, we have already gracefully shut down the SaiSwitch and its associated ports and port connectors in the SDK, so they no longer exist. As a result, we run into errors when trying to remove objects that no longer exist.

This error is somewhat expected, as per the agent's reasoning, we shouldn't even be calling the destructors.
One potential solution is to use the SaiObjects release method (https://fburl.com/code/v9rjv2qw) for objects we don't want to call the SDK for during destructor calls.

However, this would also require tweaking or suppressing a lot of additional code in the SaiManagerTable and SaiPortManager destructors (e.g., https://fburl.com/code/cceftfvu, https://fburl.com/code/uemmuxdd) to avoid touching them, leading to Credo-specific branching, which would be less elegant.

I've adopted an approach similar to the agent's. Instead of releasing the entire ensemble, I release the PhyManager from the WedgeManager during setupForWarmboot, but only for Sai Platforms (specifically Elbert with XPHYs).

Reviewed By: harshitgulati18

Differential Revision: D73395467

fbshipit-source-id: 0eb4b0754d2b1e5e9831cecb92243ee0e4841b7e