On update failure, move pre update groupIds to InputOutputState.groupIdCache
Summary:
Main reason for caching these IDs is to maintain a persistent
set of group IDs in case we have to rollback to earlier state in
case of failed updates. So rather than polluting the main EcmpResourceManager
data structures, store these in InputOutputState.groupId cache. These
will next be looked up when assigning groupIds to a new set of nhops.
NOTE: group Id persistence is mainly important for tracking merged groups,
so some of this work is in anticipation of merging algorithm.

Reviewed By: msomasundaran

Differential Revision:
D74705540

Privacy Context Container: L1125642

fbshipit-source-id: 8e4b6bb052b65665efc93b5e7574b0dc3c8d28cf