Use media type from qsfp service in prbs link tests
Summary:
We are currently using media type from agent's sw state in media type comparison in prbs test. The media type that agent has is the one qsfp service sent to it in the programInternalPhyPorts call. At this time, qsfp service doesn't know which lane is configured in what speed (that's after the programInternalPhyPorts call finishes), thus the media type is not accurate as well.

When ASIC-TransceiverSystem400GFr4 PRBS test is run on meru800bia's fabric ports, it incorrectly picks up the ports that are configured in 8x100G-FR1.

Fixing this by using the mediaType from qsfp service instead

Reviewed By: xanabani

Differential Revision:
D74707002

Privacy Context Container: L1125642

fbshipit-source-id: b060e75075bf40f1345b7c3a99b356d20dd82a09