Ignore learn event if bridgeport is not present.
Summary:
As titled.

We could receive FDB event from SDK, when the bridge port is not present. This happens due to race condition. Consider the below call sequence when we are changing the port speed.

1. Received port speed change via config.
2. Agent flushed FDB entries.
3. SDK/ASIC learnt the MAC again and sent callback.
4. Call back from SDK is enqueued in event base.
5. Agent removed bridge port from vlan.
6. Agent removed vlan.
7. Agent removed bridge port.
8. Agent removed port.
9. Agent created port with new speed.
10. Agent created vlan.
11. Agent created bridge port.
12. Agent add bridgeport to vlan.

If you look at the above sequence, if we process the enqueued learn event anytime between #7 and #12, we will crash, because SDK doesn't have the bridge port and we are trying to query it's attribute(port). SDk will return ITEM_NOT_FOUND.

Hence, we are ignoring the SDK error message.

Reviewed By: zechengh09

Differential Revision:
D74672820

Privacy Context Container: L1125642

fbshipit-source-id: 3436b9591042741a05cf76a6426a534d9428a223