Sync port link state
Summary:
As titled, post port re-create.

**Problem:**
After port speed is changed, `fboss2 show port` shows port as UP when it's really is DOWN.

**Root Cause:**
When we want to change port speed, we remove the current port and create it again with new speed. As port speed attribute is a create only attribute.

Due to below call sequence, we end up in the problem state.

Initial State: Port speed is 200G and it is UP (as neighbor is also in 200G)

1. Agent receives config to change the port speed to 400G.
2. Agent sets admin state as DOWN on port by calling SDK API.
3. Agent removes/cleans up port related objects and data structures (bridge port, VLAN, PFC, Stats, etc.,)
4. SDK calls back with link state DOWN event.
5. Agent removes port via SAI API.
6. Agent processes link state DOWN event from SDK and throws a log saying port is NOT found.
7. Agent re-creates port with new speed.
8. Agen also re-creates all necessary objects back (VLAN, bridgeport, etc.,)

Now if we see, the newly created port link state is same link state of previous incarnation. Because we missed the link DOWN event sent by SDK at #4, because by the time we processes it, we had deleted the port.

**Solution:**
Query link state of port on re-create.

TLDR;

Note: We also thought, if we put a requirement on vendor SDK to send port link state every time we create a port (irrespective of if its state changed), this problem would be solved without us doing anything. But then, we would have to request this from all vendors. Hence thought of handling it ourselves.

**Background**:

Before delving into the details of the solution, let's understand how link states are synced from SDK to wedge_agent.

1. Agent registers a call back `__glinkStateChangedNotification`with SDK to call on link state change.
2. SDK calls the callback method `__glinkStateChangedNotification` on change in link state.
3. This method calls `linkStateChangedCallbackTopHalf` which also runs in SDK thread context.
4. `linkStateChangedCallbackTopHalf` method enqueues the event in linkStateBottomHalfEventBase_ which will be processed later in FbossEventBaseThread.  This is done so that we can return the callback from SDK immediately.
5. When this even is processed
5. So, even if link state toggles `n` numbers of times, an event for each toggle will be enqueues to `linkStateBottomHalfEventBase_` in order.
6. So, both liveliness and correctness or the link state sync is achieved.

**Solution in Detail:**

Since we decided to query port link state ourselves, potentially we could create a race condition and end up in in-correct state.

Consider the below sequence.
1. Agent created the port with new speed.
2. Agent queried the port link state, let's call it is in S1.
3. Port link state changed from S1 --> S2.
4. SDK called the callback to set the state to S2.
5. Callback from SDK is  enqueued to set the state to S2. Let;s call it E1
6. We enqueue the event to set the state to S1. Let's call it E2
7. FbossEventBaseThread dequeues and processes E1, so the state in software wwitch is set to S2.
8. FbossEventBaseThread dequeues and processes E2, so the state in software switch is set to S1. Which is in-correct.

So, overcome this, the solution we implemented the following:

- Instead of enquing the request to set the link state of port in software switch, we enqueue the request to query the SDK and also set the link state of port in software switch.

In other words, we just enqueue, an event in `linkStateBottomHalfEventBase_` with portID, let;s call it, `linkStateChangedBottomHalf` event. When this event is run by fbossEventBaseThread, it will querry the SDK for link state and then do the call back to software switch to set the port link state.

**Proof of Correctness**

We will prove by contradiction that FBOSS software switch port link state will always be in sync with ASIC port link state provided below two claims hold true.

Claim 1: Events enqueued in ` linkStateBottomHalfEventBase_ ` are executed in the order they are enqueued.
Claim 2: SDK calls back for every change in link state.

**Proof of correctness by contradiction:**

Let's assume port link state in SDK is S1 and port link state in FBOSS software switch is S2, where S1 != S2.

This means, last event we processed from linkStateBottomHalfEventBase_ set the FBOSS software switch state state to S2.  We have two cases.

Case 1:
If this last event is the one we enqueued as a result for creating port, then this event was executed before the link state changed from S2 --> S1. Since we did not receive SDK call back event for change in state (S2-->S1) it voilates Claim2.

Case 2:
If this last event was enqueued as a result of SDK call back, and since we set the software switch state as S2, call back was for state change S1-->S2, once again we did not receive callback for state change from S2-->S1, which violates Claim2.

Hence proved!

Reviewed By: zechengh09

Differential Revision:
D74672821

Privacy Context Container: L1125642

fbshipit-source-id: 64ddf7656abc0ae7f3996b2afdfdaa8bce62fe25