Modify bulk programming logic
Summary:
As titled. But this is no-op as getBulkProgramCount() is NOT modified yet.
```
X = ASIC NDP scale number.
B' = 99% of X
loop = X/B';
B = B' * loop;
while(loop--):
    bulk add B' entries

Now add  more entries until we hit error.
```

In the above logic as long as B` > 50% of X, loop will always be 1. So, we can simplify the code, but leaving it as is, so that can return lower % from BulkProgramCount()

Reviewed By: jasmeetbagga

Differential Revision: D75052742

fbshipit-source-id: 6ab1346d6e6fe47ba52c698fe1048e3a25187238