Add api to setOverrideEcmpMode for rib routes
Summary:
Similar to setClassID implementation, provide a
API to set override ECMP mode in rib routes. This is to
keep the representation consistent. Further along, this
then also schedules a FIB update to ensure the pointers
are consistent.
This is a bit inefficient since the HW already has the
new override ecmp mode. Ideally we could have done
this consolidation in FIB update fn itself and update the
RIB.
Will benchmark in subsequent diffs and see if  this needs
to be improved.

Differential Revision: D75313626

fbshipit-source-id: 9e1305ad77166a5d4cd0650ab46bd62e9e8582ea