Temporarily disable state comparison in SaiSwitch stateChangedImpl
Summary:
This changed was introduced in D74686501. Since a vector of deltas is passed in, the previous applied state and current delta is compared if it was correctly applied.

This comparison is causing a few benchmarks to exceed perf limits.
https://www.internalfb.com/sandcastle/workflow/2179742219655264092

Temporarily disable the comparison. This will not affect the behavior because SaiSwitch still receives a single delta in the vector and removing the comparison simply restores to prior behavior.

Reviewed By: msomasundaran

Differential Revision: D75534664

fbshipit-source-id: 8d1d73884c2dc31e1071601df763fc9cc9af7ea1