Add uniqueDevAddress to ModbusDeviceInfo for duplicated addresses in Rackmon
Summary:
As we develop functionality to allow multiple devices on an RMC's
buses to share the same address, we need something to help differentiate them -
this diff introduces a uniqueDevAddress which will always identify one device.

Internally, we have the concept of a port which is an 8-bit unsigned integer that
identifies the /dev/tty device used to connect to the device - which is defined
in our configs.  This is the high byte in uniqueDevAddress.
Since clients wanted one integer to uniquely identify a device,
we  have allowed devAddress in all requests to be what it is today (without a
port) or take a uniqueDevAddress.

This will not impact any of the existing fleet where there are no duplicated
addresses. As we move forward, clients should start using uniqueDevAddress
to prepare for GB300.

Reviewed By: amithash, kawmarco

Differential Revision: D73698775

fbshipit-source-id: efd27889e2cf0ff451497750e573fbbd3af0c990