Reading MPI flags
Summary:
This diff adds support for reading and processing PAM4 MPI (Modulation Performance Indicator) alarm and warning flags from CMIS-compliant optical transceivers.

The implementation:

1.  Adds a new field `lanePam4MPIFlags` to the VdmPerfMonitorPortSideStats struct in transceiver.thrift to store the MPI flag values
2.  Defines new CMIS fields for PAGE_UPPER2CH and PAM4_MPI_ALARMS in cmis.thrift
3.  Implements a new function `fillVdmPerfMonitorPam4AlarmData` in CmisModule.cpp to read and process the alarm and warning flags from VDM page. The location of MPI flags are described in the comments in code.

1) from meta's Guidance for MPI Diagnostic Feature Implementation for tyhe vendors:

 {F1978591340}  {F1978591341}

After discussing with Absar, it was agreed that vendors implementing MPI should also implement corresponding alarms/flags. But to be on the safe side, we also verify that the threshold ID associated with MPI values is set to 34 as recommended, guaranteeing that vendors have defined thresholds and will raise alarms/warnings for MPI.

 {F1978591348}

Reviewed By: birdsoup

Differential Revision: D74558031

fbshipit-source-id: 832eed72a78662e16be5cfd4e4e3f8893bf54826