Fix max ucmp/ecmp member check on th4
Summary:
The logic for computing ecmp member exceed condition uses a multiple of 4 with ecmp group member count to account for the space needed in hw table for encoding the member weight. In addition, we had set the limit as 1/4th of the hw limit which is 64k. This double accounting for weight encoding is not required.

Based on scale test results, we start seeing hw full errors with 60k members while test passes with 59k. Setting the limit to 56k to provide a buffer.

Reviewed By: zechengh09

Differential Revision: D75662181

fbshipit-source-id: 2f67aa2ab0fd73ecade15bfe8f2a6c4879adc727