Optimize creating new deltas for route enqueues
Summary:
When programming a new route on ecmp limit reached we have
to decide whether this should be grafted onto the current
delta or a new delta must be enqueued. New delta only comes
into play when  we are merging groups at ECMP limit. When doing
so we will
a. Create a merged group
b. Move all routes that point to any members of this merged
group to this new group.
In transient state this creates 2 extra groups -
New merged ECMP group
Group that was merged into the new ECMP group
This is exactly the make before break buffer we have.
We need this update to be applied fully before we process
more routes. Since multiple routes maybe migrated to new
merged ECMP group and we cannot rely on ordered processing
of routes to take care of not tipping over the limit.

New delta creation *does not apply* to spillover to backup ECMP type
since, on ecmp demand exceeding, we always spillover the new
incoming route to backup ecmp type (new group only has one
route pointing to it and there is nothing cheaper than a
group pointed to by one route).

Reviewed By: msomasundaran

Differential Revision:
D75658443

Privacy Context Container: L1125642

fbshipit-source-id: be78c44ea23aca08ebe924d5edeb48364f087b23