Migrate Link Snapshot Logging to Dedicated Logs
Summary:
**Problem**
This change aims to improve the logging mechanism for snapshots in our services. Currently, we have been logging snapshots in the service log files (e.g., wedge_agent/qsfp_service), which has several inherent issues. The log lines can be quite large for snapshots, and we are currently using absurd overrides to accommodate that. Moreover, these large logs take up a significant amount of space in the log files.

**Solution**
** Not using the agent asynclogger anymore**
To address these issues, we are migrating the snapshot logs to dedicated snapshot logs. We have utilized the Folly::AsyncFileWriter class, which logs to persistent logs through an intermediate buffer.

**Implementation**
In terms of implementation, we have created a singleton instance of the AsyncFileWriter for each service. AsyncFileWriter (a subclass of AsyncLogWriter) uses a two-buffer mechanism that allows for efficient and concurrent logging by alternating between two buffers, one being written to by the IO thread while the other is filled with new log messages by other threads.

This singleton instance is then distributed as shared pointers to all instances of SnapshotManagers.

It's worth noting that we are not removing the old logging to the service file immediately. Instead, we will keep both old and new logs until everything is stable for 1/2 weeks . Once we have verified that the new logging mechanism is working correctly with log rotation, we will make the flip in the snapshotlib to read the new logs and stop the old logging altogether.

For snapshot agent, we re keeping the logging as it is. The snapshot service log is already dedicated for snapshotting, creating another log does not make sense.

Related diff: D67804252

Reviewed By: birdsoup

Differential Revision: D67814533

fbshipit-source-id: 19f3ea6c86c425eb6272c2b1b9dc6b0ae0a88266