Set remote intf routes to drop if neighbor is unresolved
Summary:
The issue was observed in S519958:

We install unicast routes on RDSW for DTSW's inband IP, which is pointing to remote interfaces on EDSW. When port flaps on EDSW, neighbor entry is removed and hence the unicast route is updated to point to CPU port and assign classID 2.

For routes that are pointing to remote interfaces, we should set the route to drop instead of CPU, since there will be no NDP probe for remote devices.

To fix the issue, have ManagedRouteNextHop carry the information of whether resolved intf is local or remote. If it's pointing to remote, when the next hop goes away we should set the route to DROP rather than to CPU.

Reviewed By: jasmeetbagga

Differential Revision:
D74767157

Privacy Context Container: L1125642

fbshipit-source-id: ba25530695d3c27987c9435432cda169b67952ae