disambiguate attributes for early drop and ecn wred objects
Summary:
this is a temporary change to make warm boot work. once SAI api is fixed to correctly return values and support separate early drop and ecn thresholds it will have to reverted

at this time sdk returns early drop attributes as well for ECN objects in chenab. this breaks warm boot as set attribute tuple is different in fboss than in sai.

Differential Revision: D75894682

fbshipit-source-id: f59ffeb43409b0b01e0ce1e6edab28c0d5b41a2e