fix roll back tests
Summary:
rollback tests now work with warm boot support (store reload) supported

updating route prefixes used in test as these prefixes are overlapping and conflicting with rif addresses configured for chenab.

as a result 100.0.0.0/24 is actually a rif address and not a route with ecmp next hops.

same goes for 200.0.0.0/24, 100::/64 and 200::/64

Differential Revision: D76009930

fbshipit-source-id: 00f64ef83c24ba35c253b5f2554974a8292982b9