ASIC revision attribute support
Summary:
### ASIC revision attribute support
#### Summary

This diff adds support for the ASIC revision attribute to the Switch API.

The `AttributeAsicRevision` struct is added to `SaiSwitchTraits` in `SwitchApi.h`. This struct provides a way to retrieve the ASIC revision attribute.

Implementations for the `AttributeAsicRevision` struct are added to `SwitchApi.cpp` files for Tajo, OSS, and BCM.

In the BCM implementation, the `SAI_SWITCH_ATTR_ASIC_REVISION` attribute is used if the BRCM SAI SDK version is DNX GTE 13.0. Otherwise, the attribute is not supported.

Reviewed By: shri-khare

Differential Revision: D76063303

fbshipit-source-id: de21b502c090d04775b5b140c8dcc25bc140bdd4