Fix Bug for Manual Sync Configo Diff When Run on Master
Summary:
In the case that a manual sync is needed when configerator platform mapping data falls out of sync with fbcode, devs can run `buck2 run mode/opt fbcode//fboss/lib/platform_mapping_v2/facebook:gen -- --make-configerator-changes`.

However, there is a minor bug that leads the generated Configo diff to have a dependency on the current checked out diff, regardless of what it is. This is most relevant when running the manual sync from master – the generated Configo diff will then have a dependency on the most recently checked out fbcode diff on master.

This diff adds a new flag that only enables dependency setting if `add-fbcode-dependency` is enabled. This flag is added to the `platform_mapping_v2_verify_and_configo_sync` command, so it is only enabled in on-diff jobs.

Reviewed By: xanabani

Differential Revision: D75811280

fbshipit-source-id: 5c8738c17c6058031479ac1a954cf9d9cd1a7236