Try reopening watchdog if it's not opened
Summary:
# Summary
In case the file is invalid(doesn't exist)- open is gonna return -1 making the optional variable having the value. Resulted in us always trying to write to invalid file handler.
> On success, open(), openat(), and creat() return the new file descriptor (a nonnegative integer).  On error, -1 is returned and errno is set to indicate the error.

This diff considers the fact that open returns -1 on error and it tries to reopen the watchdog file if the handler is invalid.

Reviewed By: Scott8440

Differential Revision: D75337411

fbshipit-source-id: 97176a2b01293cb6f519c6c4d138ca9cbed12e4e