Address roundtrip test error due to test name too long
Summary:
This test is added to know bad for round trip due to file name too long ( > 255)

```
2025/06/03 22:08:28.299|10923|MainThread|D|paramiko_utils: gzip: /tmp/8283044a83aa4f93b556fa841d28f48d_leaba_1_42_8_1_42_8_gibraltar_mono_roundtrip_preprod_2_trunk_2_preprod_warm_boot_for_warm_boot_0_agentemptytest_checkinit_agenthwptptcprovisiontests_agenthwptptcprovisiontests_verifyptptcdelayrequestonports_twentyfiveg_.out: File name too long
```

A better fix is to have some "ondiff" check once a new test was introduced (how to do it is not clear yet). Created T227306955 to track.

For now, do a simple change here to avoid this long file name.

With the fix, the longest file name for roundtrip should be ~ 240

Differential Revision: D76471678

fbshipit-source-id: 2bf59c96f8f868d6a3cc7e8cdd52390d41bd7bfd