Support warmboot for nhop-group from array to map type
Summary:
Until the old array type for nhop-group adapterHostKey gets deprecated, SAI warm boot state would have 2 versions.
Use the nhop-group-temporary map going forward.

After migrating all SAI switches, the temporary map will not be written to and the existing nhop-group would start using the map (instead of the array). This is possible because of the change made in this diff to follyDynamicToAdapterHostKey where the dynamic type is derived automatically.

Reviewed By: msomasundaran

Differential Revision: D76327744

fbshipit-source-id: 05bd5bbcc75e1033edd1d1edc2d36c892604effb