Ensure we reset the state machine on the hard reset thrift call
Summary:
When the thrift call to hard reset is made, we go through the safety checks (isSafeToRemove) and not move the transceiver to NOT_PRESENT state. We still hard reset the transceiver. Thus the transceiver was reset, but the state machine never resets. This leaves the module in an inconsistent state, causing various side effects such as port not recovering correctly, diagnostics  not getting initialized etc as seen with the FEC tails missing post auto remediation in S526319.

Thrift API to hard reset is meant to forcefully reset a transceiver, no matter what the current state of it is. We use this thrift API in auto-remediation triggered by Self Healing.

Added a forceRemoveTransceiver attribute in the state machine, which is set only when the thrift call to reset transceiver is made.

Reviewed By: xanabani

Differential Revision:
D76649951

Privacy Context Container: L1125642

fbshipit-source-id: 45d7779c21884683c400374aeee211d98031598a