sdk to not load and unload kmod every time switch is created and removed
Summary:
In Chenab, by default SDK loads kmod on create switch and unloads kmod on remove switch

this is not how FBOSS works and expect kmod to be loaded and unloaded. The following change signals sdk to not load kmod and unload kmod every time create and remove switch is invoked.

Differential Revision:
D76763099

Privacy Context Container: L1125642

fbshipit-source-id: 2391e357df5a6dcdaaefdde3a64d309cd95e0902