Fix shadow variable warning in CmdShowInterfaceCountersMKA.h
Summary:
## Instructions about RACER Diffs:
This diff fixes a 'Shadow Variable Warning' issue identified by Quality Insight from [Monetization codehub](https://fburl.com/quality/wmkbc0si):
 - Accept and ship the diff (racer does not do auto land)
 - [**Recommended**] Commandeer and land this diff.

If you are happy with the changes, please accept and ship, or commandeer it if minor edits are needed.
If you are not happy with the changes, please comment on the diff with clear actions and send it back to the author. Racer will pick it up and regenerate.
If you really feel the Racer is not helping with this change (alas, some complex changes are hard for AI) feel free to abandon this diff.

This diff was pre-created by Racer AI agent for your convenience on top of T227995241. More details about the program see this [post](https://fb.workplace.com/groups/1177219810659335/permalink/1198503531864296/). For questions or suggestions please post in [RACER Feedback and Q&A](https://fb.workplace.com/groups/2477474979269093) group.
## Summary:
Fixed a shadow variable warning in CmdShowInterfaceCountersMKA.h by renaming the inner 'table' variable to 'portStatsTable' to avoid shadowing the outer 'table' variable declared in the printOutput method.
 ---
> Generated by [RACER](https://www.internalfb.com/wiki/RACER_(Risk-Aware_Code_Editing_and_Refactoring)/), powered by [Confucius](https://www.internalfb.com/wiki/Confucius/Analect/Shared_Analects/Confucius_Code_Assist_(CCA)/)
[Session](https://www.internalfb.com/confucius?session_id=a83cec80-4be3-11f0-b607-af7a4631928e&tab=Chat), [Trace](https://www.internalfb.com/confucius?session_id=a83cec80-4be3-11f0-b607-af7a4631928e&tab=Trace)

Differential Revision: D76867907

fbshipit-source-id: 76ce41571c006c4fda8c8ac557f86f9bf9831bfd