Supporting Speed change test for Elbert macsec platforms
Summary:
We encountered  QSFP crashes during speed changes for elbert macsec devices. Upon investigation, we discovered that the changePortImpl implementation for phy SaiPortManager was buggy. Specifically, we were attempting to create new ports and port connectors where existing ones already existed.

To resolve this issue, I consulted with the agent team and decided to modify the changePortImpl method to remove the ports before adding them back.

Additionally, we adjusted the timeout in the setup for runSpeedChangeTest from 1 minute to 300 seconds. Otherwise, we are seeing ports not coming up within 1 min. That is kind of expected. This is consistent with our approach in the link test cold boot setup stage (https://fburl.com/code/m83m1enz), where we also use a 300-second timeout. The reason for this adjustment is that changing speed configurations introduces additional overhead for the QSFP service due to xPHY reprogramming, which involves removing and re-adding SAI entities. This should be allowed similar time  like a pure coldboot.

Reviewed By: birdsoup

Differential Revision: D77003940

fbshipit-source-id: edac3988961587fc96d2f28297665f51d626d9ea