Convert PlatformMapping from Unique Ptr to Shared Ptr
Summary:
As of right now, one of `TransceiverManager`'s constructor arguments is `unique_ptr<PlatformMapping>`, which is used for iterating through all platform ports, all chips (e.g. for initiating a state machine for each transceiver). However, since we discussed the creation of `TransceiverManager` and `PortManager` as two separate objects, both will need a pointer to this `const PlatformMapping` object.

Towards this effort, this diff:
- Converts all `TransceiverManager` and `WedgeManager` constructors to take in `shared_ptr<PlatformMapping>` instead of `unique_ptr<PlatformMapping>`.
- Adds const qualifiers wherever possible to this shared_ptr.

The alternative would be to rely on `TransceiverManager`'s `getPlatformMapping()` ([link](https://fburl.com/code/kz8c0h3k)) function, but we should avoid using raw pointers wherever possible. Currently, `getPlatformMapping()` function is used for initialization of platform-specific PhyManager (e.g. [link](https://fburl.com/code/3k0b60op)), but we should aim to change this later as a follow-up task.

The next diff will move the creation of this `shared_ptr` outside of the platform-specific `WedgeManager`constructors, so that it can be passed as an argument to both `TransceiverManager` and `PortManager` down the line.

Reviewed By: harshitgulati18

Differential Revision: D75731132

fbshipit-source-id: 61960c8f17e4c5cbbb1caa8cdace29b336ac9b3c